refactor(ruborist): extract placement.rs and file.rs from builder
The remaining two cuts from the builder.rs structure review: the graph
placement helpers (create/reuse/place tail, chain_err, the demand
handoff) move to resolver/placement.rs next to edges.rs, and the
feature-gated file: resolver moves to resolver/file.rs mirroring
git.rs/http.rs. Pure moves; builder re-exports the public surface so
the lib.rs facade and the demand driver are unchanged. builder.rs is
now the resolution config + entry points + spec router (1164 lines,
~480 of them tests; 1561 before the series).
